backend/requirements.txt
semantic-kernel==1.28.0
fastapi==0.115.12
uvicorn[standard]==0.34.0
azure-identity==1.21.0
azure-search-documents==11.6.0
azure-cosmos==4.9.0
opentelemetry-sdk==1.32.0
opentelemetry-instrumentation-fastapi==0.53b0
azure-monitor-opentelemetry==1.6.4
pydantic-settings==2.9.1
python-dotenv==1.1.0
